requirements-minimal.txt
Flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
websocket-client>=1.8.0
